[Behat][Promotion] Changed test promotion factory logic
