+++ etc/behat/services/contexts.xml
@@ -94,9 +94,9 @@
 
         <service id="sylius.behat.context.setup.promotion" class="%sylius.behat.context.setup.promotion.class%" scope="scenario">
             <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
-            <argument type="service" id="sylius.repository.promotion" container="symfony" />
-            <argument type="service" id="sylius.repository.promotion_action" container="symfony" />
+            <argument type="service" id="sylius.factory.promotion_action" container="symfony" />
             <argument type="service" id="sylius.test.factory.promotion" container="symfony" />
+            <argument type="service" id="sylius.repository.promotion" container="symfony" />
             <argument type="service" id="doctrine.orm.entity_manager" container="symfony" />
             <tag name="sylius.behat.context" />
         </service>
